Fix some warnings on Windows (#2560)
1. Enable warning "4503" # Decorated name length exceeded.
2. Enable warning "4146" # unary minus operator applied to unsigned type.
3. Enable float64 support for the Softmax operator
4. Enable compliance checks for Windows x86 32bits build
5. Use TryBatchParallelFor to replace some fallback code in mlas pooling.cc
6. Fix Android CI pipeline.

+++ onnxruntime/core/session/onnxruntime_c_api.cc
@@ -910,7 +910,8 @@ struct CallGetValueImpl {
     const auto* tensor_data = tensor.Data<TensorElemType>();
     OrtStatus* st = OrtApis::CreateTensorAsOrtValue(allocator, shape.GetDims().data(), shape.NumDimensions(),
                                                     onnxruntime::utils::GetONNXTensorElementDataType<TensorElemType>(), out);
-    return st ? st : PopulateTensorWithData(*out, tensor_data, shape.Size(), sizeof(TensorElemType));
+    //TODO: check overflow before doing static_cast
+    return st ? st : PopulateTensorWithData(*out, tensor_data, static_cast<size_t>(shape.Size()), sizeof(TensorElemType));
   }
 };
 
@@ -1079,7 +1080,8 @@ static OrtStatus* OrtCreateValueImplSeqHelperTensor(const Tensor& tensor,
     return st;
   }
 
-  size_t num_elems = tensor.Shape().Size();
+  //TODO: check the cast below
+  size_t num_elems = static_cast<size_t>(tensor.Shape().Size());
   auto* out_data = out.MutableData<TensorElemType>();
   for (size_t i = 0; i < num_elems; ++i) {
     *out_data++ = *data++;
